restore stable version

+++ src/main/java/com/example/demo/services/CommentService.java
@@ -3,7 +3,6 @@ package com.example.demo.services;
 import com.example.demo.domain.Comment;
 import com.example.demo.domain.Post;
 import com.example.demo.repositories.CommentRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Update;
@@ -16,12 +15,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
     private final CommentRepository commentRepo;
+    private final MongoTemplate mongoTemplate;
 
-    @Autowired
-    private MongoTemplate mongoTemplate;
-
-    public CommentService(CommentRepository commentRepo) {
+    public CommentService(CommentRepository commentRepo, MongoTemplate mongoTemplate) {
         this.commentRepo = commentRepo;
+        this.mongoTemplate = mongoTemplate;
     }
 
     @Transactional
@@ -37,7 +35,7 @@ public class CommentService {
         commentRepo.insert(comment);
 
         mongoTemplate.update(Post.class)
-                .matching(Criteria.where("postId").is(postId))
+                .matching(Criteria.where("_id").is(postId))
                 .apply(new Update().push("comments").value(comment))
                 .first();
 
@@ -45,7 +43,18 @@ public class CommentService {
     }
 
     public void deleteComment(String id){
-        commentRepo.deleteById(id);
+        Comment comment = commentRepo.findById(id);
+        if (comment != null) {
+            String postId = comment.getParentId();
+            commentRepo.deleteById(id);
+            
+            if (postId != null) {
+                mongoTemplate.update(Post.class)
+                        .matching(Criteria.where("_id").is(postId))
+                        .apply(new Update().pull("comments", comment))
+                        .first();
+            }
+        }
     }
 
 }
